remove caregiver_mulesoft_config_v2 feature toggle and usages
