+++ config/features.yml
@@ -147,9 +147,6 @@ features:
   caregiver_use_rails_logging_over_sentry:
     actor_type: user
     description: Use Rails for logging instead of Sentry
-  caregiver_mulesoft_config_v2:
-    actor_type: user
-    description: Use refactored mulesoft configuration
   cerner_non_eligible_sis_enabled:
     actor_type: user
     description: Enables Sign in Service for cerner authentication
