+++ lib/carma/client/mule_soft_auth_token_client.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'carma/client/mule_soft_auth_token_configuration'
-require 'carma/client/mule_soft_auth_token_configuration_v2'
 
 module CARMA
   module Client
@@ -12,6 +11,8 @@ module CARMA
       GRANT_TYPE = 'client_credentials'
       SCOPE = 'DTCWriteResource'
 
+      configuration MuleSoftAuthTokenConfiguration
+
       class GetAuthTokenError < StandardError; end
 
       def new_bearer_token
@@ -19,8 +20,7 @@ module CARMA
           response = perform(:post,
                              auth_token_path,
                              params,
-                             token_headers,
-                             *(Flipper.enabled?(:caregiver_mulesoft_config_v2) ? [] : [{ timeout: config.timeout }]))
+                             token_headers)
 
           return JSON.parse(response.body)['access_token'] if response.status == 200
 
@@ -29,14 +29,6 @@ module CARMA
       end
 
       private
-
-      def config
-        if Flipper.enabled?(:caregiver_mulesoft_config_v2)
-          MuleSoftAuthTokenConfigurationV2.instance
-        else
-          MuleSoftAuthTokenConfiguration.instance
-        end
-      end
 
       def params
         URI.encode_www_form({
