+++ lib/carma/client/mule_soft_configuration.rb
@@ -3,17 +3,11 @@
 module CARMA
   module Client
     class MuleSoftConfiguration < Common::Client::Configuration::REST
-      def base_request_headers
-        super.merge({
-                      client_id: settings.client_id,
-                      client_secret: settings.client_secret
-                    })
-      end
-
       def connection
-        Faraday.new(base_path) do |conn|
+        Faraday.new(base_path, headers: base_request_headers) do |conn|
           conn.use(:breakers, service_name:)
           conn.request :instrumentation, name: service_name
+          conn.options.timeout = timeout
           conn.adapter Faraday.default_adapter
         end
       end
@@ -23,9 +17,9 @@ module CARMA
       end
 
       # @return [Integer] Value given by configuration key `form_10_10cg.carma.mulesoft.timeout`
-      # setting. Defaults to 10 if unset.
+      # setting. Defaults to 600 if unset.
       def timeout
-        settings.key?(:timeout) ? settings.timeout : 10
+        settings.key?(:timeout) ? settings.timeout : 600
       end
 
       # @return [Config::Options]
@@ -33,11 +27,21 @@ module CARMA
         Settings.form_10_10cg.carma.mulesoft
       end
 
+      def base_request_headers
+        super.merge(
+          'Authorization' => "Bearer #{bearer_token}"
+        )
+      end
+
       private
 
       # @return [String]
       def base_path
         "#{settings.host}/va-carma-caregiver-papi/api/"
+      end
+
+      def bearer_token
+        @bearer_token ||= CARMA::Client::MuleSoftAuthTokenClient.new.new_bearer_token
       end
     end
   end
