+++ spec/lib/carma/client/mule_soft_auth_token_client_spec.rb
@@ -53,83 +53,32 @@ describe CARMA::Client::MuleSoftAuthTokenClient do
 
     let(:mock_token_response) { Faraday::Response.new(response_body:, status: 200) }
 
-    context ':caregiver_mulesoft_config_v2 toggle disabled' do
-      before do
-        allow(Flipper).to receive(:enabled?).with(:caregiver_mulesoft_config_v2).and_return(false)
-      end
+    context 'successfully gets token' do
+      it 'calls perform with expected params' do
+        expect(client).to receive(:perform)
+          .with(
+            :post,
+            config.settings.auth_token_path,
+            token_params, token_headers
+          )
+          .and_return(mock_token_response)
 
-      it 'uses MuleSoftAuthTokenConfiguration config' do
-        allow(client).to receive(:config).and_call_original
-        expect(client.send(:config)).to be_a(CARMA::Client::MuleSoftAuthTokenConfiguration)
-      end
-
-      context 'successfully gets token' do
-        it 'calls perform with expected params' do
-          expect(client).to receive(:perform)
-            .with(
-              :post,
-              config.settings.auth_token_path,
-              token_params, token_headers, options
-            )
-            .and_return(mock_token_response)
-
-          expect(subject).to eq token
-        end
-      end
-
-      context 'error getting token' do
-        let(:mock_error_token_response) { Faraday::Response.new(response_body: { sad: true }, status: 400) }
-
-        it 'raises error' do
-          expect(client).to receive(:perform)
-            .with(:post, config.settings.auth_token_path, token_params, token_headers, options)
-            .and_return(mock_error_token_response)
-
-          expect do
-            subject
-          end.to raise_error(CARMA::Client::MuleSoftAuthTokenClient::GetAuthTokenError,
-                             "Response: #{mock_error_token_response}")
-        end
+        expect(subject).to eq token
       end
     end
 
-    context ':caregiver_mulesoft_config_v2 toggle enabled' do
-      before do
-        allow(Flipper).to receive(:enabled?).with(:caregiver_mulesoft_config_v2).and_return(true)
-      end
+    context 'error getting token' do
+      let(:mock_error_token_response) { Faraday::Response.new(response_body: { sad: true }, status: 400) }
 
-      it 'uses MuleSoftAuthTokenConfigurationV2 config' do
-        allow(client).to receive(:config).and_call_original
-        expect(client.send(:config)).to be_a(CARMA::Client::MuleSoftAuthTokenConfigurationV2)
-      end
+      it 'raises error' do
+        expect(client).to receive(:perform)
+          .with(:post, config.settings.auth_token_path, token_params, token_headers)
+          .and_return(mock_error_token_response)
 
-      context 'successfully gets token' do
-        it 'calls perform with expected params' do
-          expect(client).to receive(:perform)
-            .with(
-              :post,
-              config.settings.auth_token_path,
-              token_params, token_headers
-            )
-            .and_return(mock_token_response)
-
-          expect(subject).to eq token
-        end
-      end
-
-      context 'error getting token' do
-        let(:mock_error_token_response) { Faraday::Response.new(response_body: { sad: true }, status: 400) }
-
-        it 'raises error' do
-          expect(client).to receive(:perform)
-            .with(:post, config.settings.auth_token_path, token_params, token_headers)
-            .and_return(mock_error_token_response)
-
-          expect do
-            subject
-          end.to raise_error(CARMA::Client::MuleSoftAuthTokenClient::GetAuthTokenError,
-                             "Response: #{mock_error_token_response}")
-        end
+        expect do
+          subject
+        end.to raise_error(CARMA::Client::MuleSoftAuthTokenClient::GetAuthTokenError,
+                           "Response: #{mock_error_token_response}")
       end
     end
   end
